+++ app/renderer/pages/errorPage/index.js
@@ -73,7 +73,7 @@ class ErrorPage extends React.Component {
                   settings <a href='javascript:void(0);' onClick={() => onNavToSettingGeneralPage()} style={{ color: colors.N0 }}>here</a> and
                   restart the application again, or
                   feel free
-                  to <a href='javascript:void(0);' onClick={() => this.props.handleClick ? this.props.handleClick() : openExternalLink('mailto:support@runanode.io')} style={{ color: colors.N0 }}>contact us</a> if
+                  to <a href='javascript:void(0);' onClick={() => this.props.handleContactUsClick ? this.props.handleContactUsClick() : openExternalLink('mailto:support@runanode.io')} style={{ color: colors.N0 }}>contact us</a> if
                   the problem persists
                 </Note>
                 {isDevOrDebugProd && (